global: move to new code structure

NOTE: beware, patch amended by Tibor to use new file names
