+++ invenio/testsuite/tester_utils_datacite.py
@@ -26,7 +26,7 @@ the tests.
 
 from invenio.flaskshell import *
 
-from invenio import dataciteutils
+from invenio.utils import datacite as dataciteutils
 from invenio import config
 from invenio.testsuite import make_test_suite, run_test_suite
 import random
